model + path

+++ lib/core/eval_visdrone.py
@@ -51,7 +51,7 @@ def compute_success_error(gt_center, result_center):
 
 def get_result_bb(arch, seq):
     result_path = join(arch, seq + '.txt')
-    temp = np.loadtxt(result_path, delimiter=',').astype(np.float)
+    temp = np.loadtxt(result_path, delimiter=',').astype(np.float32)
     return np.array(temp)
 
 
@@ -77,7 +77,7 @@ def eval_auc(dataset='VISDRONVAL', result_path='./result/', tracker_reg='S*', st
     for i in range(n_seq):
 
         seq = seqs[i]
-        gt_rect = np.array(annos[seq]['gt_rect']).astype(np.float)
+        gt_rect = np.array(annos[seq]['gt_rect']).astype(np.float32)
         gt_center = convert_bb_to_center(gt_rect)
         for j in range(len(trackers)):
             tracker = trackers[j]
@@ -114,7 +114,7 @@ def eval_vis_tune(result_path, dataset='VISDRONEVAL'):
 
     for i in range(n_seq):
         seq = seqs[i]
-        gt_rect = np.array(annos[seq]['gt_rect']).astype(np.float)
+        gt_rect = np.array(annos[seq]['gt_rect']).astype(np.float32)
         gt_center = convert_bb_to_center(gt_rect)
         bb = get_result_bb(result_path, seq)
         center = convert_bb_to_center(bb)
